[Editorial] Replace italicization of word with quotes.
I originally read the italics as marking usage of a French word.
However, a more careful reading suggests that the italics mark the
emphasis of using a technical term. Merriam-Webster indicates that
petite entered English in the 18th century, but the first recorded
usage as a specific size for women's clothes is 1926, the year this
book was published. I believe this is the meaning that is being used
here, and the phrasing of this sentence indicates that it should
therefore be in quotes.